models/tdims/requirements.txt
numpy>=1.26.1, <2.0.0
pandas>=1.5.3, <2.2.0
rdkit-pypi>=2022.9.4, <2023.9.6
scikit-learn>=1.2.0, <1.5.0
shap>=0.42.1, <0.43.0
matplotlib>=3.10.1, <3.11.0